==> notebooks/3_feature_engineering.ipynb <==
from pyspark.sql import SparkSession
from pyspark.ml.feature import Tokenizer, StopWordsRemover, HashingTF, IDF, StringIndexer
from pyspark.ml import Pipeline

# Stop any existing Spark context
try:
    spark.stop()
except:
    pass

# Optimized for i5 11th gen + 16GB RAM
spark = SparkSession.builder \
    .appName("Music_Classifier") \
    .master("local[2]") \
    .config("spark.driver.memory", "4g") \
    .config("spark.executor.memory", "2g") \
    .config("spark.sql.execution.arrow.pyspark.enabled", "false") \
    .getOrCreate()

df_train = spark.read.csv("./notebook_data/Mendeley_cleaned_train.csv", header=True, inferSchema=True)

# Encode genre labels
genre_indexer = StringIndexer(inputCol="genre", outputCol="label")

# Tokenize and clean lyrics
tokenizer = Tokenizer(inputCol="lyrics", outputCol="words")
remover = StopWordsRemover(inputCol="words", outputCol="filtered_words")

# TF-IDF
hashingTF = HashingTF(inputCol="filtered_words", outputCol="raw_features", numFeatures=10000)
idf = IDF(inputCol="raw_features", outputCol="features")

df_train.show(5)

# Build pipeline
pipeline = Pipeline(stages=[genre_indexer, tokenizer, remover, hashingTF, idf])
pipeline_model = pipeline.fit(df_train)
df_transformed = pipeline_model.transform(df_train)

df_transformed.show(5)

# Save pipeline
pipeline_model.write().overwrite().save("./notebook_data/feature_pipeline_lyrics_only")
